frontend: Split MainInfoSection's header into its own component
This will be useful for a split of the different DetailsGrid we're
preparing.
